requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
psycopg2-binary